requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.5.0
anthropic==0.7.8
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
structlog==23.2.0
prometheus-client==0.19.0
phonenumbers==8.13.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
